the main activity is divided into files by function, the character container class has been removed
