+++ app/src/main/java/com/example/android_course/Character.kt
@@ -1,23 +1,7 @@
 package com.example.android_course
 
-class Character (
-    name: String,
-    resourcesLink: String,
-    welcomeMessage: String)
-{
-    private var _name : String = name
-    private var _resourcesLink : String = resourcesLink
-    private var _welcomeMessage = welcomeMessage
-
-    fun getName() : String {
-        return _name
-    }
-
-    fun getLink() : String {
-        return _resourcesLink
-    }
-
-    fun getMessage() : String {
-        return _welcomeMessage
-    }
-}
+data class Character(
+    val name: String,
+    val imageLink: String?,
+    val welcomeMessage: String
+)
