+++ app/src/main/java/com/example/android_course/Constants.kt
@@ -1,23 +1,23 @@
 package com.example.android_course
 
-var MARVER_LOGO_LINK = "https://iili.io/JMnuvbp.png"
+val MARVEL_LOGO_LINK = "https://iili.io/JMnuvbp.png"
 
-var CHARACTERS = listOf(
+val CHARACTERS = listOf(
     Character(
         name = "Deadpool",
-        resourcesLink = "https://iili.io/JMnAfIV.png",
+        imageLink = "https://iili.io/JMnAfIV.png",
         welcomeMessage = "I'm deadpool, baby!"
     ),
     Character(
         name ="Iron-Man",
-        resourcesLink = "https://iili.io/JMnuDI2.png",
+        imageLink = "https://iili.io/JMnuDI2.png",
         welcomeMessage = "I am Iron Man."
     ),
     Character(
         name = "Spider-Man",
-        resourcesLink = "https://iili.io/JMnuyB9.png",
+        imageLink = "https://iili.io/JMnuyB9.png",
         welcomeMessage = "Your friendly neighborhood Spider-Man."
     )
 )
 
-var CHARACTERS_CONTAINER = CharactersContainer()
+
